Moved topPlayerlist
