+++ Varo - 1.7 - 1.15/src/de/cuuky/varo/game/leaderboard/TopScoreList.java
@@ -1,4 +1,4 @@
-package de.cuuky.varo.scoreboard;
+package de.cuuky.varo.game.leaderboard;
 
 import java.util.ArrayList;
 import java.util.Collections;
