+++ Varo - 1.7 - 1.15/src/de/cuuky/varo/scoreboard/tablist/TablistHandler.java
@@ -1,0 +1,8 @@
+package de.cuuky.varo.scoreboard.tablist;
+
+public class TablistHandler {
+	
+	public TablistHandler() {
+		
+	}
+}
